Updated SQL Table and inputed dummy data so the Controller could be tested. Adjusted SQL DAO files

+++ java/src/main/java/com/techelevator/dao/CollectionDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.techelevator.model.Collection;
 
 public interface CollectionDAO {
-	Collection getCollectionById(int collectionId);
-	List<Collection> getAllCollectionsByUserId(int userId);
-	void updateCollection(int userId);
+	Collection getCollectionById(Long collectionId);
+	List<Collection> getAllCollectionsByUserId(Long userId);
+	void updateCollection(Long userId);
 }
